chore: moved operations and functions to separate file
